Fix missing requirement for uri
When loading configuration from an option file (using
OpenSearch::CLI#load, which rely on OptionParser#load) and `uri` is not
loaded, an undefined method exception is raised when an `url` option is
parsed, but that exception is captured by the OptionParser logic
resulting in the option file to be completely ignored.

While #load return false to indicate the failure, consumers of the
library are unlikely to check it because end users may or may not rely
on option files.

Unconditionally require uri to avoid this issue.